Use TPL Dataflow
Rewrite the code to use System.Threading.Tasks.Dataflow.

Finding applicable files is still handled sequentially on the main thread, as it was before. Reading to produce the set of lines is now handled by a TransformManyBlock; finding matches within a line is handled by a TransformBlock; and outputing the results is handled by an ActionBlock.

The two transform blocks are allowed to operate in parallel. The intent is to maximize the throughput of file reading and matching operations.

Since the results still need to be output sequentially, the ActionBlock for outputing the results maintains its degress of parallelism at the default (1).